requirements.txt
streamlit
requests
pandas
plotly
jinja2
python-dateutil
playwright
openai
